# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Nodelo Backend API",
    "description": "RESTful API for Nodelo website backend. Includes contact form submissions, authentication, real-time chat with file uploads, and admin management endpoints. Features Socket.io for real-time messaging and Cloudinary for file storage.",
    "version": "1.0.0",
    "contact": {
      "name": "Nodelo Team",
      "email": "hello@nodelo.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5050",
      "description": "Development server"
    },
    {
      "url": "https://api.nodelo.com",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/api/contacts": {
      "post": {
        "tags": ["Contacts"],
        "summary": "Create contact submission",
        "description": "Submit a new contact form. This is a public endpoint that sends emails to admin and user.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "message"],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100,
                    "example": "John Doe"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "john.doe@example.com"
                  },
                  "company": {
                    "type": "string",
                    "maxLength": 200,
                    "example": "Acme Inc",
                    "nullable": true
                  },
                  "projectType": {
                    "type": "string",
                    "enum": ["starter-site", "mvp-web-app", "dapp-contract", "support-ops", "other"],
                    "example": "mvp-web-app",
                    "nullable": true
                  },
                  "budget": {
                    "type": "string",
                    "enum": ["under-1k", "1k-5k", "5k-10k", "10k-25k", "25k-plus"],
                    "example": "5k-10k",
                    "nullable": true
                  },
                  "message": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 5000,
                    "example": "I'm interested in building a web application for my business."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Contact submission created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Contact form submitted successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "example": "507f1f77bcf86cd799439011"
                        },
                        "name": {
                          "type": "string",
                          "example": "John Doe"
                        },
                        "email": {
                          "type": "string",
                          "example": "john.doe@example.com"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Contacts"],
        "summary": "Get all contacts",
        "description": "Retrieve all contact submissions with pagination. Requires authentication.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Filter by status",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["new", "replied", "archived"]
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number (default: 1)",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Items per page (default: 20)",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of contacts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Contact"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/chats": {
      "post": {
        "tags": ["Chats"],
        "summary": "Create new chat",
        "description": "Create a new chat conversation. Requires authentication.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["userName", "userEmail"],
                "properties": {
                  "userName": {
                    "type": "string",
                    "example": "John Doe"
                  },
                  "userEmail": {
                    "type": "string",
                    "format": "email",
                    "example": "john@example.com"
                  },
                  "adminId": {
                    "type": "string",
                    "example": "507f1f77bcf86cd799439012",
                    "nullable": true,
                    "description": "Admin ID to assign. Defaults to authenticated admin."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Chat created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Chat"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Chats"],
        "summary": "Get all chats",
        "description": "Retrieve all chat conversations with pagination and filtering. Requires authentication.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Filter by chat status",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["active", "resolved", "archived"]
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Items per page",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of chats",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Chat"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /api/auth/login endpoint. Include in Authorization header as: Bearer <token>"
      }
    },
    "schemas": {
      "Contact": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "507f1f77bcf86cd799439011"
          },
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "john.doe@example.com"
          },
          "company": {
            "type": "string",
            "nullable": true,
            "example": "Acme Inc"
          },
          "projectType": {
            "type": "string",
            "enum": ["starter-site", "mvp-web-app", "dapp-contract", "support-ops", "other"],
            "nullable": true,
            "example": "mvp-web-app"
          },
          "budget": {
            "type": "string",
            "enum": ["under-1k", "1k-5k", "5k-10k", "10k-25k", "25k-plus"],
            "nullable": true,
            "example": "5k-10k"
          },
          "message": {
            "type": "string",
            "example": "I'm interested in building a web application for my business."
          },
          "status": {
            "type": "string",
            "enum": ["new", "replied", "archived"],
            "default": "new",
            "example": "new"
          },
          "emailSent": {
            "type": "boolean",
            "default": false,
            "example": true
          },
          "chatId": {
            "type": "string",
            "nullable": true,
            "example": "507f1f77bcf86cd799439013",
            "description": "Linked chat conversation ID"
          },
          "repliedBy": {
            "type": "string",
            "nullable": true,
            "example": "507f1f77bcf86cd799439014",
            "description": "Admin user ID who replied"
          },
          "replyCount": {
            "type": "integer",
            "default": 0,
            "example": 2,
            "description": "Number of admin replies"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00.000Z"
          }
        },
        "required": ["name", "email", "message", "status"]
      },
...
      "Pagination": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "example": 20
          },
          "total": {
            "type": "integer",
            "example": 100
          },
          "pages": {
            "type": "integer",
            "example": 5
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "Error message description"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true,
            "example": ["Email is required", "Password must be at least 6 characters"]
          },
          "message": {
            "type": "string",
            "nullable": true,
            "example": "Additional error details (development only)"
          }
        }
      },
      "Chat": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "507f1f77bcf86cd799439011"
          },
          "contactId": {
            "type": "string",
            "nullable": true,
            "example": "507f1f77bcf86cd799439012",
            "description": "Linked contact submission ID"
          },
          "userId": {
            "type": "string",
            "example": "user@example.com",
            "description": "User identifier (usually email)"
          },
          "userName": {
            "type": "string",
            "example": "John Doe"
          },
          "userEmail": {
            "type": "string",
            "format": "email",
            "example": "john@example.com"
          },
          "adminId": {
            "type": "string",
            "nullable": true,
            "example": "507f1f77bcf86cd799439013",
            "description": "Assigned admin user ID"
          },
          "adminName": {
            "type": "string",
            "nullable": true,
            "example": "Admin User",
            "description": "Name of the admin handling this chat"
          },
          "status": {
            "type": "string",
            "enum": ["active", "resolved", "archived"],
            "default": "active",
            "example": "active"
          },
          "lastMessageAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "lastMessageBy": {
            "type": "string",
            "enum": ["user", "admin"],
            "example": "user"
          },
          "unreadCount": {
            "type": "object",
            "properties": {
              "user": {
                "type": "integer",
                "example": 2
              },
              "admin": {
                "type": "integer",
                "example": 5
              }
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00.000Z"
          }
        }
      },
...
    }
  }
}